+++ src/lib.rs
@@ -7,8 +7,9 @@ use hyper::{body, Body, Request, Response, Server};
 use log::{debug, error, info, trace, warn};
 use serde::de::DeserializeOwned;
 use std::sync::Arc;
+mod prometheus_metric;
 mod render_to_prometheus;
-pub use render_to_prometheus::PrometheusMetric;
+pub use prometheus_metric::PrometheusMetric;
 mod metric_type;
 mod prometheus_instance;
 pub use metric_type::MetricType;
